(PC-6770) Make ProfileContainer and Separator reusable
